Replace 'client modules' and '@jar' with metadata rules
Note that this is a firs step that reproduces exactly the same
dependencies as the build had before the change. The intention why
dependencies where excluded or changed before is lost for most cases.
Some changes are most likely not even intended.
Each rule needs to be checked and it needs to be be clarified why it
is required.